+++ waspc/cli/exe/Main.hs
@@ -38,6 +38,7 @@ import Wasp.Cli.Command.Test (test)
 import Wasp.Cli.Command.TsConfigSetup (tsConfigSetup)
 import Wasp.Cli.Command.Uninstall (uninstall)
 import Wasp.Cli.Command.WaspLS (runWaspLS)
+import Wasp.Cli.Common (getInstallationCommand, getPackagingMode)
 import Wasp.Cli.Message (cliSendMessage)
 import Wasp.Cli.Terminal (title)
 import qualified Wasp.Message as Message
@@ -209,15 +210,17 @@ printUsage =
 
 printVersion :: IO ()
 printVersion = do
+  installCommand <- getInstallationCommand <$> getPackagingMode
+
   putStrLn $
     unlines
       [ show waspVersion,
         "",
         "If you wish to install/switch to the latest version of Wasp, do:",
-        "  curl -sSL https://get.wasp.sh/installer.sh | sh -s",
+        indent 2 $ installCommand Nothing,
         "",
         "If you want specific x.y.z version of Wasp, do:",
-        "  curl -sSL https://get.wasp.sh/installer.sh | sh -s -- -v x.y.z",
+        indent 2 $ installCommand $ Just "x.y.z",
         "",
         "Check https://github.com/wasp-lang/wasp/releases for the list of valid versions, including the latest one."
       ]