+++ waspc/tests/AppSpec/ValidTest.hs
@@ -99,10 +99,9 @@ spec_AppSpecValid = do
                              [ "Your Wasp version does not match the app's requirements.",
                                "You are running Wasp " ++ show WV.waspVersion ++ ".",
                                "This app requires Wasp ^" ++ show incompatibleWaspVersion ++ ".",
-                               "To install specific version of Wasp, do:",
-                               "  curl -sSL https://get.wasp.sh/installer.sh | sh -s -- -v x.y.z",
-                               "where x.y.z is your desired version.",
-                               "Check https://github.com/wasp-lang/wasp/releases for the list of valid versions."
+                               "To get instuctions on how to install the correct Wasp version, run:",
+                               "  wasp version",
+                               "and follow the instructions there."
                              ]
                        ]